Trait web3_closest_provider::ClosestWeb3Provider

pub trait ClosestWeb3Provider {
    // Required methods
    fn init(urls: Vec<String>, checking_interval: Duration) -> Self;
    fn is_ready(&self) -> bool;
    fn destroy(&self);
    fn get_fastest_provider(&self) -> String;
    fn wait_until_ready(&self) -> impl Future<Output = ()> + Send;
}

Defines methods for interacting with a Web3 provider balancer. This trait enables you to:

Initialize a balancer with a list of Web3 provider URLs.

Check if the balancer is ready to provide the fastest provider.

Stop the balancer and clear its data.

Get the URL of the provider with the fastest response time.

Wait until the balancer is ready to provide the fastest provider.

This trait is useful for applications that need to dynamically select the fastest Web3 provider based on response time, ensuring optimal performance for your Web3 operations.

Required Methods

fn init(urls: Vec<String>, checking_interval: Duration) -> Self

Initializes the provider balancer with a list of URLs.

Example

use web3_closest_provider::{ClosestWeb3Provider, ClosestWeb3RpcProviderSelector};
use std::time::Duration;

#[tokio::main]
async fn main() {
    let providers = vec![
        "https://mainnet.infura.io/v3/your_api_key".to_string(),
        "https://rpc.ankr.com/eth".to_string(),
        "https://api.mycryptoapi.com/v1/eth".to_string(),
    ];

    let balancer = ClosestWeb3RpcProviderSelector::init(providers.clone(), Duration::from_secs(10));
}

Arguments

urls - A vector of URLs for the Web3 providers.

checking_interval - The interval at which the balancer checks the response times of the providers.

fn is_ready(&self) -> bool

Checks if the balancer is ready to provide the fastest provider.

Example

use web3_closest_provider::{ClosestWeb3Provider, ClosestWeb3RpcProviderSelector};
use std::time::Duration;

#[tokio::main]
async fn main() {
    let providers = vec![
        "https://mainnet.infura.io/v3/your_api_key".to_string(),
        "https://rpc.ankr.com/eth".to_string(),
        "https://api.mycryptoapi.com/v1/eth".to_string(),
    ];

    let balancer = ClosestWeb3RpcProviderSelector::init(providers.clone(), Duration::from_secs(10));

    if balancer.is_ready() {
        println!("Balancer is ready to use!");
    } else {
        balancer.wait_until_ready().await;    
        println!("Balancer is ready to use!");
    }
}

Returns

true if the balancer is ready; false otherwise.

fn destroy(&self)

Stops the balancer and clears its data.

Example

use web3_closest_provider::{ClosestWeb3Provider, ClosestWeb3RpcProviderSelector};
use std::time::Duration;

#[tokio::main]
async fn main() {
    let providers = vec![
        "https://mainnet.infura.io/v3/your_api_key".to_string(),
        "https://rpc.ankr.com/eth".to_string(),
        "https://api.mycryptoapi.com/v1/eth".to_string(),
    ];

    let balancer = ClosestWeb3RpcProviderSelector::init(providers.clone(), Duration::from_secs(10));

    balancer.destroy(); // **This step is essential!**
}

fn get_fastest_provider(&self) -> String

Returns the URL of the provider with the fastest response time.

Example

use web3_closest_provider::{ClosestWeb3Provider, ClosestWeb3RpcProviderSelector};
use std::time::Duration;

#[tokio::main]
async fn main() {
    let providers = vec![
        "https://mainnet.infura.io/v3/your_api_key".to_string(),
        "https://rpc.ankr.com/eth".to_string(),
        "https://api.mycryptoapi.com/v1/eth".to_string(),
    ];

    let balancer = ClosestWeb3RpcProviderSelector::init(providers.clone(), Duration::from_secs(10));

    balancer.wait_until_ready().await;   
    let fastest_provider = balancer.get_fastest_provider();
    println!("Fastest provider: {}", fastest_provider);
}

// ... use the fastest provider for your Web3 operations ...

Returns

The URL of the provider with the fastest response time.

Panics

This function will panic if the hashmap containing response times is empty.

fn wait_until_ready(&self) -> impl Future<Output = ()> + Send

Waits until the balancer is ready to provide the fastest provider.

Example

use web3_closest_provider::{ClosestWeb3Provider, ClosestWeb3RpcProviderSelector};
use std::time::Duration;

#[tokio::main]
async fn main() {
    let providers = vec![
        "https://mainnet.infura.io/v3/your_api_key".to_string(),
        "https://rpc.ankr.com/eth".to_string(),
        "https://api.mycryptoapi.com/v1/eth".to_string(),
    ];

    let balancer = ClosestWeb3RpcProviderSelector::init(providers.clone(), Duration::from_secs(10));

    balancer.wait_until_ready().await;
    println!("Balancer is ready to use!");
}

Object Safety

This trait is not object safe.